+++ src/docs/changelog.md
@@ -13,10 +13,12 @@ Changelog                           {#changelog}
 - #362 removed iwidget::getKeyboardFocusWidget. It was not used
 - #306 using ChaiScript for scripting
 - added offset to camera component
-- #457 TODO
+- #457 adding script support to Mica
 - fixed recent projects order
 - fixed edit mode issue after project creation
 - fixed window closing on all client events
+- clear dialogs on close to prevent duplicate widgets leading to mem leak
+- fixed issue with widget type print
 
 0.45.0
 ------